requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
PyYAML>=6.0
pytesseract>=0.3.10
Pillow>=10.0.0
colorlog>=6.7.0
